# apply_stylesheet optimizations