--- a/include/litehtml/html_tag.h
+++ b/include/litehtml/html_tag.h
@@ -72,6 +72,7 @@ namespace litehtml
 
 		int					select(const css_selector& selector, bool apply_pseudo = true) override;
 		int					select(const css_element_selector& selector, bool apply_pseudo = true) override;
+		int					select_pseudoclass(const css_attribute_selector& attr);
 
 		elements_vector		select_all(const string& selector) override;
 		elements_vector		select_all(const css_selector& selector) override;