--- a/src/css_selector.cpp
+++ b/src/css_selector.cpp
@@ -21,6 +21,17 @@ void litehtml::css_element_selector::parse( const string& txt )
 			attribute.attribute	= "class";
 			m_attrs.push_back(attribute);
 			el_end = pos;
+		} else if(txt[el_end] == '#')
+		{
+			css_attribute_selector attribute;
+
+			string::size_type pos = txt.find_first_of(".#[:", el_end + 1);
+			attribute.val		= txt.substr(el_end + 1, pos - el_end - 1);
+			litehtml::lcase(attribute.val);
+			attribute.condition	= select_id;
+			attribute.attribute	= "id";
+			m_attrs.push_back(attribute);
+			el_end = pos;
 		} else if(txt[el_end] == ':')
 		{
 			css_attribute_selector attribute;
@@ -64,16 +75,6 @@ void litehtml::css_element_selector::parse( const string& txt )
 				m_attrs.push_back(attribute);
 				el_end = pos;
 			}
-		} else if(txt[el_end] == '#')
-		{
-			css_attribute_selector attribute;
-
-			string::size_type pos = txt.find_first_of(".#[:", el_end + 1);
-			attribute.val		= txt.substr(el_end + 1, pos - el_end - 1);
-			attribute.condition	= select_equal;
-			attribute.attribute	= "id";
-			m_attrs.push_back(attribute);
-			el_end = pos;
 		} else if(txt[el_end] == '[')
 		{
 			css_attribute_selector attribute;