--- a/src/css_selector.cpp
+++ b/src/css_selector.cpp
@@ -2,6 +2,46 @@
 #include "css_selector.h"
 #include "document.h"
 
+void litehtml::css_element_selector::parse_nth_child_params(const string& param, int& num, int& off)
+{
+	if (param == "odd")
+	{
+		num = 2;
+		off = 1;
+	}
+	else if (param == "even")
+	{
+		num = 2;
+		off = 0;
+	}
+	else
+	{
+		string_vector tokens;
+		split_string(param, tokens, " n", "n");
+
+		string s_num;
+		string s_off;
+
+		string s_int;
+		for (const auto& token : tokens)
+		{
+			if (token == "n")
+			{
+				s_num = s_int;
+				s_int.clear();
+			}
+			else
+			{
+				s_int += token;
+			}
+		}
+		s_off = s_int;
+
+		num = atoi(s_num.c_str());
+		off = atoi(s_off.c_str());
+	}
+}
+
 void litehtml::css_element_selector::parse( const string& txt )
 {
 	string::size_type el_end = txt.find_first_of(".#[:");
@@ -41,38 +81,55 @@ void litehtml::css_element_selector::parse( const string& txt )
 
 			if(txt[el_end + 1] == ':')
 			{
+				attribute.type = select_pseudo_element;
 				string::size_type pos = txt.find_first_of(".#[:", el_end + 2);
-				attribute.val		= txt.substr(el_end + 2, pos - el_end - 2);
-				attribute.type	= select_pseudo_element;
-				litehtml::lcase(attribute.val);
+				string name		= txt.substr(el_end + 2, pos - el_end - 2);
+				litehtml::lcase(name);
+				attribute.name = _id(name);
 				m_attrs.push_back(attribute);
 				el_end = pos;
 			} else
 			{
 				string::size_type pos = txt.find_first_of(".#[:(", el_end + 1);
-				if(pos != string::npos && txt.at(pos) == '(')
-				{
-					pos = find_close_bracket(txt, pos);
-					if(pos != string::npos)
-					{
-						pos++;
-					}
-				}
-				if(pos != string::npos)
-				{
-					attribute.val		= txt.substr(el_end + 1, pos - el_end - 1);
-				} else
-				{
-					attribute.val		= txt.substr(el_end + 1);
-				}
-				litehtml::lcase(attribute.val);
-				if(attribute.val == "after" || attribute.val == "before")
+				string name = txt.substr(el_end + 1, pos - el_end - 1);
+				lcase(name);
+				attribute.name = _id(name);
+				if(attribute.name == _after_ || attribute.name == _before_)
 				{
 					attribute.type	= select_pseudo_element;
 				} else
 				{
 					attribute.type	= select_pseudo_class;
 				}
+
+				string val;
+				if(pos != string::npos && txt.at(pos) == '(')
+				{
+					auto end = find_close_bracket(txt, pos);
+					val = txt.substr(pos + 1, end - pos - 1);
+					if (end != string::npos) pos = end + 1;
+				}
+
+				switch (attribute.name)
+				{
+				case _nth_child_:
+				case _nth_of_type_:
+				case _nth_last_child_:
+				case _nth_last_of_type_:
+					lcase(val);
+					parse_nth_child_params(val, attribute.a, attribute.b);
+					break;
+				case _not_:
+					attribute.sel = std::make_shared<css_element_selector>();
+					attribute.sel->parse(val);
+					break;
+				case _lang_:
+					trim(val);
+					lcase(val);
+					attribute.val = val;
+					break;
+				}
+
 				m_attrs.push_back(attribute);
 				el_end = pos;
 			}