Results on Obama wiki:
```
Before:
> apply_stylesheet: 2848 ms
document::createFromString: 3711 ms

After:
> apply_stylesheet: 1025 ms
document::createFromString: 1856 ms
```

<hr>

## Tests

I think current litehtml tests are not very useful. The only useful thing they do is to test if litehtml actually compiles.
I am bothered by them because I had to fix compilation errors in them multiple times already, and those fixes were massive.
With this PR I again found myself in this situation. I fixed several tests in cssTest.cpp, but just deleted others.
How testing parsing `:default` and `:checked` and a bunch of other parameterless pseudo-classes any helpful? They hit exactly the same code path.

Until now I just fixed tests without too much thinking, but now I took a closer look, and some of them don't even test anything.
For example, `StyleAddProperty` in cssTest.cpp doesn't have any checks, it just calls `style.add_property` repetitively.
And I had to fix that function not once, but twice already, all 81 lines. This is so annoying, completely useless work.

Is it OK if next time I delete tests that don't test anything and tests that check the same code path?