@tordex 2022-11-16
> Is it OK if next time I delete tests that don't test anything and tests that check the same code path?

Yes, it OK.